feat: did a better job ignoring the db connector
feat: factored helper functions out of server
