+++ projectMetaData.json
@@ -270,7 +270,7 @@
         "text-color": "black"
       }
     },
-    "downloads": "0",
+    "downloads": 1,
     "palette": {
       "1": {
         "color": "#90908c",
